connect conversation controller to inputprovider (#1023)
* Update ConversationEvents.cs

* chore: bump version to 0.1.1005

* fix conversation ids

* chore: bump version to 0.1.1006

* Update HubSimpleConversationTrigger.cs

* chore: bump version to 0.1.1007

* try to fix weather race condition

* chore: bump version to 0.1.1008

* connect conversation controller to inputprovider

* chore: bump version to 0.1.1009

---------

+++ Assets/TurnrootFramework/Gameplay/Brain/Components/Input/Battle/BatttleInputController/BattleInputControllerBrain.cs
@@ -115,7 +115,7 @@ namespace Turnroot.Gameplay.Brain
             {
                 var inputVec = _inputActions.Navigate.ReadValue<Vector2>();
                 var camAngle = Brain.cameraBrain.CurrentAngle;
-                var steps = (((int)camAngle % 360) + 360) % 360 / 90;
+                var steps = ((camAngle % 360) + 360) % 360 / 90;
                 var direction = RotateVectorBy90StepsCW(inputVec, steps);
                 if (direction.magnitude > 0.1f)
                 {
